udemy_langgraph/reflexion_agent/pyproject.toml
[project]
name = "reflexion-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10, <3.11"
dependencies = [
    "langchain-community>=0.3.17",
    "langchain-core>=0.3.34",
    "langchain-openai>=0.3.4",
    "langgraph>=0.2.70",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
]